+++ lib/app/data/datasources/coin_datasource.dart
@@ -1,0 +1,5 @@
+import 'package:everest_crypto/app/domain/entities/coin_entity.dart';
+
+abstract class CoinDatasource {
+  Future<List<CoinEntity>> getAllCoins();
+}
