+++ lib/app/domain/entities/coin_entity.dart
@@ -1,0 +1,18 @@
+import 'dart:convert';
+
+class CoinEntity {
+  final String id;
+  final String name;
+  final String image;
+  final String symbol;
+  final String latest;
+  final String amount;
+  CoinEntity({
+    required this.id,
+    required this.name,
+    required this.image,
+    required this.symbol,
+    required this.latest,
+    required this.amount,
+  });
+}
