+++ lib/app/domain/repositories/coin_repository.dart
@@ -1,0 +1,5 @@
+import 'package:everest_crypto/app/domain/entities/coin_entity.dart';
+
+abstract class CoinRepository {
+  Future<List<CoinEntity>> getAllCoins();
+}
