+++ lib/app/presenter/controllers/visible_controller.dart
