--- lib/features/app/data/datasources/coin_datasource.dart
@@ -1,5 +1,0 @@
-import 'package:everest_crypto/features/details/domain/entities/coin_entity.dart';
-
-abstract class CoinDatasource {
-  Future<List<CoinEntity>> getAllCoins();
-}
