--- lib/features/app/domain/repositories/coin_repository.dart
@@ -1,5 +1,0 @@
-import 'package:everest_crypto/features/details/domain/entities/coin_entity.dart';
-
-abstract class CoinRepository {
-  Future<List<CoinEntity>> getAllCoins();
-}
