--- lib/features/app/presenter/ui/portfolio/view/portfolio_page.dart
@@ -1,53 +1,0 @@
-import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../model/coin.dart';
-import '../repositories/coin_repository.dart';
-import '../controllers/visible_controller.dart';
-import '../widgets/coin_list.dart';
-import '../widgets/wallet_details.dart';
-
-class PortfolioPage extends HookConsumerWidget {
-  const PortfolioPage({Key? key}) : super(key: key);
-
-  static const route = '/portfolio';
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final visible = ref.watch(visibleNotifierProvider);
-    CoinRepository repository = CoinRepository(Dio());
-    List<Coin> coinList = repository.getAllCoins();
-
-    return Scaffold(
-      backgroundColor: const Color.fromRGBO(255, 255, 255, 1),
-      body: SafeArea(
-        child: Column(
-          children: [
-            const SizedBox(
-              height: 8,
-            ),
-            Padding(
-              padding: const EdgeInsets.all(
-                18.0,
-              ),
-              child: WalletDetails(
-                  visible: visible,
-                  repository: repository,
-                  changeVisibility: () {
-                    ref.watch(visibleNotifierProvider.state).state = !visible;
-                  }),
-            ),
-            CoinList(
-              coinList: coinList,
-              visible: visible,
-            ),
-            const SizedBox(
-              height: 10,
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
